Project 4 Update
Updated with project4.html and both scripts to validate the contact forms in the project4 page and the contact page. Both forms now work (sorta) with a simulated redirect from the page (no php or server side handling yet) using a different (but essentially identical in function) script for each. Minor css styling changes for the website as a whole. Fixed the margin-bottom on the sidebar element which was appearing much longer than it needed to.